+++ requirements.php
@@ -12,7 +12,6 @@
 // you may need to adjust this path to the correct Yii framework path
 
 define("APP_CONFIG_DIR",        realpath(__DIR__ . '/../config'));
-define("APP_RUNTIME_DIR",       realpath(__DIR__ . '/../runtime'));
 define("ROOT_DIR",              dirname(dirname(__DIR__)));
 
 $frameworkPath = ROOT_DIR . '/vendor/yiisoft/yii2';