+++ CHANGELOG.md
@@ -3,20 +3,29 @@ CHANGELOG
 
 4.0.0.pre
 ---------------
+ * Using paulzi/yii2-auto-tree
  * Removed skeeks\cms\base\CheckComponent
  * Removed skeeks\cms\traits\ValidateRulesTrait
  * Removed skeeks\cms\Config
  * Change code folder /src
  * Url changes
  
-3.2.3.0.pre
+3.2.4.pre
 -----------------
+ * Filter rewrite sql
+ 
+3.2.3
+-----------------
+ * Fancybox return v2.*
+ * Fixed admin url
+ * Fixed fancybox
  * Fixed login url
  * Uploads files for new not saved models (change \skeeks\cms\models\behaviors\HasStorageFile)
 
 3.2.2.2
 -----------------
  * Fixed fancybox
+ * Url changes
  * ContentElementsCmsWidget new setting isJoinTreeMap
  * allowDeselect PropertyTypeElement
  * Documentation about working with the url